Reset the heap configuration file
Normally, the heap configuration file is set in device/qcom/common
and the default configuration is xhdpi.
This is not needed for 8x10 as it as a hdpi based device,
and setting it to xhdpi based heap configuration only increases
memory utilization.

Override default configuration with the appropriate heap configuration.

Change-Id: I9924378587b6864e27dbe39a64162702983006f1