keylayout: Fix Camera hard button press issue
Issue: Camera keycode is not passed to framework when Camera hard
key is pressed, resulting in image capture not working when camera
key pressed.

Fix: WAKE_DROPPED Flag is removed in L-framework code so the
parsing of gpio-keys.kl file at bootup fails. Hence when Camera
key is pressed the corresponding scan code map is not available
and unknown keycode map is returned to framework. Removing
WAKE_DROPPED flag from the gpio-keys.kl file ensures parsing and
correct scan code to be returned to framework.

Change-Id: Iacd218cf2959fb51d25e8c4112612379f794fdd8
CRs-Fixed: 710844